XRT-1320 Build support added for AzureSphere SDK 22.02 and API 12

+++ src/c/scheduler.c
@@ -78,7 +78,7 @@ static uint64_t getTimeAsUInt64 (void)
 {
   struct timespec ts;
   clock_gettime (CLOCK_REALTIME, &ts);
-  return (uint64_t) ts.tv_sec * IOT_BILLION + ts.tv_nsec;
+  return (uint64_t) ((uint64_t) ts.tv_sec * IOT_BILLION) + (uint64_t) ts.tv_nsec;
 }
 
 /* Scheduler thread function */
